--- reducer/index.js
@@ -1,10 +1,0 @@
-import { combineReducers } from 'redux';
-import user from './user';
-import cart from './cart';
-
-const reducer = combineReducers({
-  user,
-  cart
-});
-
-export default reducer;
