+++ src/Cart.js
@@ -1,17 +1,16 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { getProducts } from '../store';
-import { setCartThunks, updateThunks } from '../reducer/cart';
+import { getProducts, setCartThunks } from '../store';
 
 class _Cart extends React.Component {
-  constructor({ cart }) {
+  constructor() {
     super();
     this.state = {
-      quantity: cart.quantity
+      // quantity: cart.quantity
     }
     this.destroy = this.destroy.bind(this);
-    this.update = this.update.bind(this);
+    // this.update = this.update.bind(this);
   }
   async destroy(id) {
     await this.props.destroy(id);
@@ -25,22 +24,21 @@ class _Cart extends React.Component {
     console.log(this.props)
     return (
       <div>
-        {/* Cart Items: { products.filter(product => product.cartId === cart.productId).length} */}
+        Cart Items: { products.filter(product => product.id === cart.productId).length}
         <ul>
           {
-           products.map( product =>
-            // cart.map( c => products.find( product => product.id === c.productId 
-            //  ?
+           products.map( product => product.id === cart.productId
+             ?
             <div key={product.id} id='flex'>
                 <Link to={`/products/${product.id}`} activeclassname="active"><li key='img'><img src ={product.imageURL}></img></li></Link>
                 <Link to={`/products/${product.id}`} activeclassname="active"><h1 key='name'>{product.name}</h1></Link>
                 <li key='genre'>{product.genre}</li>
                 <li key='price'>{product.price}</li>
-                <button onClick = { updateCart({...cart, quantity: this.state.quantity + 1 }) } >+</button> {/* increase quantity of item in cart */}
-                <button onClick = { updateCart({...cart, quantity: this.state.quantity - 1 }) } >-</button> {/* decrease quantity of item in cart */}
+                {/* <button onClick = { updateCart({...cart, quantity: this.state.quantity + 1 }) } >+</button> increase quantity of item in cart */}
+                {/* <button onClick = { updateCart({...cart, quantity: this.state.quantity - 1 }) } >-</button> decrease quantity of item in cart */}
                 <button onClick = { destroy }>Delete</button>
-           </div> )
-            // : 'Nothing in cart'))
+           </div> 
+            : 'Nothing in cart')
           }
         </ul>
       </div>
@@ -58,7 +56,7 @@ const mapStateToProps = state => {
 const mapDispatchToProps = {
   getProducts: getProducts,
   setCart: setCartThunks,
-  updateCart: updateThunks
+  // updateCart: updateCartThunks
 }
 
 const Cart = connect(mapStateToProps, mapDispatchToProps)(_Cart);
